Move UI module in a separate file
